Updated InvokeFunctionMissingnAnalyzer to be consistent with other Analyzers
