+++ UnityEngineAnalyzer/UnityEngineAnalyzer.Test/UnityEngineAnalyzer.Test.csproj
@@ -139,6 +139,7 @@
     <Compile Include="OnGUI\DoNotUseOnGUIAnalyzerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StringMethods\DoNotUseStringMethodsAnalyzerTests.cs" />
+    <Compile Include="StringMethods\InvokeFunctionMissingAnalyzerTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
